# Reduce UDP server memory usage